03 CSharp/apiAppointments/requirements.txt
wheel==0.33.6
django==2.2.8
djangorestframework==3.9.2
Pillow==2.2.1
django-pyodbc-azure==2.1.0.0
django-pyodbc==1.1.3
pyodbc==4.0.28
django-mssql-backend==2.5.0
conekta==2.5.1
gunicorn==19.9.0
exponent-server-sdk==0.3.1
django-cors-headers==3.5.0
fcm-django==0.3.4
idna==2.9
pyfcm==1.4.7
requests==2.23.0
simplejson==3.17.0
Unidecode==1.1.1
djangorestframework-api-key==2.0.0 


